Use double-precision for time tick; Force bridge build without osc

+++ source/backend/CarlaEngine.hpp
@@ -280,8 +280,8 @@ struct CARLA_API EngineTimeInfoBBT {
 
     int32_t bar;  //!< current bar
     int32_t beat; //!< current beat-within-bar
-    int32_t tick; //!< current tick-within-beat
-    double barStartTick;
+    double  tick; //!< current tick-within-beat
+    double  barStartTick;
 
     float beatsPerBar; //!< time signature "numerator"
     float beatType;    //!< time signature "denominator"
